+++ packages/compass-welcome/src/index.ts
@@ -2,7 +2,7 @@ import { registerHadronPlugin } from 'hadron-app-registry';
 import { createLoggerLocator } from '@mongodb-js/compass-logging/provider';
 import { workspacesServiceLocator } from '@mongodb-js/compass-workspaces/provider';
 import type { WorkspaceComponent } from '@mongodb-js/compass-workspaces';
-import { WelcomeModal, WelcomeTab } from './components';
+import { WelcomeModal, DesktopWelcomeTab, WebWelcomeTab } from './components';
 import { activatePlugin } from './stores';
 import { telemetryLocator } from '@mongodb-js/compass-telemetry/provider';
 
@@ -12,20 +12,28 @@ const serviceLocators = {
   workspaces: workspacesServiceLocator,
 };
 
-export const WelcomePlugin = registerHadronPlugin(
-  {
-    name: 'Welcome',
-    component: WelcomeTab,
-    activate: activatePlugin,
-  },
-  serviceLocators
-);
-
-export const WorkspaceTab: WorkspaceComponent<'Welcome'> = {
+export const DesktopWorkspaceTab: WorkspaceComponent<'Welcome'> = {
   name: 'Welcome' as const,
-  component: WelcomePlugin,
+  component: registerHadronPlugin(
+    {
+      name: 'Welcome',
+      component: DesktopWelcomeTab,
+      activate: activatePlugin,
+    },
+    serviceLocators
+  ),
+};
+
+export const WebWorkspaceTab: WorkspaceComponent<'Welcome'> = {
+  name: 'Welcome' as const,
+  component: registerHadronPlugin(
+    {
+      name: 'Welcome',
+      component: WebWelcomeTab,
+      activate: activatePlugin,
+    },
+    serviceLocators
+  ),
 };
 
 export { WelcomeModal };
-
-export default WelcomePlugin;
